requirements.txt
pushbullet.py==0.12.0
pandas==1.2.0
openpyxl
